Move span annotation parsing logic to abstract data strategy
