+++ src/dnanet/data/strategies/datasets/dataset.py
@@ -17,6 +17,11 @@ from abc import ABC, abstractmethod
 from typing import Any, Dict, List, Tuple, Literal, Mapping, Generator
 from pathlib import Path
 
+import numpy as np
+import pandas as pd
+from loguru import logger
+
+from dnanet.core import LabelCategory
 
 if typing.TYPE_CHECKING:
     from pathlib import Path
@@ -24,7 +29,7 @@ if typing.TYPE_CHECKING:
     from annotated_types import T
 
     from dnanet.core.types import PathLike
-    from dnanet.core.annotation import Annotation
+    from dnanet.core.annotation import Annotation, ScanpointAnnotation
     from dnanet.data.strategies.scaling import ScalingStrategy
 
 
@@ -149,3 +154,124 @@ class DatasetStrategy(ABC):
     @property
     def annotation_to_idx(self) -> Dict[str, int]:
         return {name: idx for idx, name in enumerate(self.get_annotation_classes())}
+
+    @classmethod
+    def _parse_span_annotation(
+            cls, span_annotations_path: Path, scaling_strategy: ScalingStrategy
+    ) -> dict[str, ScanpointAnnotation | None]:
+
+        _dye_name_to_dye_idx = {
+            'blue': 0,
+            'green': 1,
+            'yellow': 2,
+            'black': 2,
+            'red': 3,
+            'purple': 4,
+            'orange': 5,
+        }
+
+        # collect files
+        csv_files = list(span_annotations_path.rglob('*.csv'))
+        if not csv_files:
+            logger.warning('No span annotation CSV files found in {}', span_annotations_path)
+            return {}
+
+        # read csv files
+        df = pd.concat((pd.read_csv(f) for f in csv_files), ignore_index=True)
+        columns = df.columns.tolist()
+        required_columns = {'profile', 'user', 'dye', 'x0', 'x1', 'category'}
+        missing_columns = required_columns.difference(df.columns)
+        if missing_columns:
+            raise ValueError(f'Missing span annotation columns: {sorted(missing_columns)}, found columns: {sorted(columns)}')
+
+        len_before_drop = len(df)
+        df = df.dropna(subset=['profile', 'user', 'dye', 'x0', 'x1', 'category']).copy()
+        dropped = len_before_drop - len(df)
+
+
+        logger.info(
+            f'Found {len(df)} valid span annotations in {len(df["profile"].unique())} '
+            f'profiles (dropped {dropped} rows)'
+        )
+        logger.info(f'Categories found in annotations: {df["category"].unique()}')
+
+        # convert dye names to dye indices
+        df['dye_idx'] = (
+            df['dye']
+            .astype(str)
+            .str.strip()
+            .str.lower()
+            .map(_dye_name_to_dye_idx)
+        )
+        unknown_dyes = df.loc[df['dye_idx'].isna(), 'dye'].unique()
+        if len(unknown_dyes) > 0:
+            raise ValueError(f'Unknown dye values in span annotations: {unknown_dyes}')
+
+        # convert category names to indices
+        df['category_idx'] = df['category'].map(LabelCategory.display_name_to_index)
+        unknown_categories = df.loc[df['category_idx'].isna(), 'category'].unique()
+        if len(unknown_categories) > 0:
+            raise ValueError(f'Unknown category values in span annotations: {unknown_categories}')
+
+        df[['dye_idx', 'category_idx', 'x0', 'x1']] = df[
+            ['dye_idx', 'category_idx', 'x0', 'x1']
+        ].astype(int)
+
+        # create span annotations grouped by file and annotator
+        hid_file_name_to_span_annotations: dict[str, list[np.ndarray]] = {}
+        for (hid_file_name, _annotator), hid_file_df in df.groupby(['profile', 'user'], sort=False):
+            spannotation = cls._df_to_span_annotation(hid_file_df, scaling_strategy)
+            hid_file_name_to_span_annotations.setdefault(hid_file_name, []).append(spannotation)
+
+        # merge span annotations into a scanpoint annotation
+        hid_to_annotation: dict[str, ScanpointAnnotation | None] = {}
+        for hid_file_name, span_annotations in hid_file_name_to_span_annotations.items():
+            if len(span_annotations) > 1:
+                span_annotation = cls._merge_span_annotations(span_annotations, hid_file_name)
+            else:
+                span_annotation = span_annotations[0]
+
+            hid_to_annotation[hid_file_name] = cls._span_to_scanpoint_annotation(span_annotation, hid_file_name)
+
+        return hid_to_annotation
+
+    @staticmethod
+    def _df_to_span_annotation(df: pd.DataFrame, scaling_strategy: ScalingStrategy) -> np.ndarray:
+        num_dyes = scaling_strategy.kit.num_dyes
+        scanpoints = scaling_strategy.scanpoint_resolution
+        num_classes = len(LabelCategory)
+        spannotation = np.zeros((num_dyes, scanpoints, num_classes), dtype=np.int8)
+
+        for row in df.itertuples(index=False):
+            dye_idx = int(row.dye_idx)
+            category_idx = int(row.category_idx)
+            if not 0 <= dye_idx < num_dyes:
+                raise ValueError(f'Dye index {dye_idx} outside annotation shape')
+            if not 0 <= category_idx < num_classes:
+                raise ValueError(f'Category index {category_idx} outside annotation shape')
+
+            start, stop = sorted((int(row.x0), int(row.x1)))
+            start = max(0, start)
+            stop = min(scanpoints, stop)
+            if start >= stop:
+                continue
+
+            spannotation[dye_idx, start:stop, category_idx] = 1
+
+        return spannotation
+
+    @staticmethod
+    def _merge_span_annotations(spannotations: List[np.ndarray], hid_file_name: str) -> np.ndarray:
+        logger.debug(
+            f'Found multiple span annotations for {hid_file_name}. Merging by taking the first only'
+        )
+        return spannotations[0]
+
+    @staticmethod
+    def _span_to_scanpoint_annotation(span_annotation: np.ndarray, hid_file_name: str) -> ScanpointAnnotation:
+        flattened = span_annotation.argmax(axis=-1)
+
+        if np.any(span_annotation.sum(axis=-1) > 1):
+            logger.debug(f'Found overlapping annotations for {hid_file_name}, taking the lowest class index')
+
+        return ScanpointAnnotation(flattened.astype(np.int8, copy=False))
